Relocate pkg
